server/model/requirements.txt
openai==1.12.0
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.1.0
langchain-openai==0.0.2.post1
tiktoken==0.5.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
